+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.11.0-rc.26",
+  sdkVersion: "4.11.0-rc.27",
   genVersion: "2.907.0",
-  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.26 2.907.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.27 2.907.0 2.0 @orq-ai/node",
 } as const;